+++ pkg/cmd/server.go
@@ -2,6 +2,12 @@ package cmd
 
 import (
 	"fmt"
+	mlflowService "github.com/G-Research/fasttrackml/pkg/service"
+	"github.com/G-Research/fasttrackml/pkg/service/artifact"
+	"github.com/G-Research/fasttrackml/pkg/service/experiment"
+	"github.com/G-Research/fasttrackml/pkg/service/metric"
+	"github.com/G-Research/fasttrackml/pkg/service/model"
+	"github.com/G-Research/fasttrackml/pkg/service/run"
 	"net/http"
 	"os"
 	"os/signal"
@@ -23,12 +29,6 @@ import (
 	aimController "github.com/G-Research/fasttrackml/pkg/api/aim/controller"
 	mlflowAPI "github.com/G-Research/fasttrackml/pkg/api/mlflow"
 	mlflowController "github.com/G-Research/fasttrackml/pkg/api/mlflow/controller"
-	mlflowService "github.com/G-Research/fasttrackml/pkg/api/mlflow/service"
-	"github.com/G-Research/fasttrackml/pkg/api/mlflow/service/artifact"
-	"github.com/G-Research/fasttrackml/pkg/api/mlflow/service/experiment"
-	"github.com/G-Research/fasttrackml/pkg/api/mlflow/service/metric"
-	"github.com/G-Research/fasttrackml/pkg/api/mlflow/service/model"
-	"github.com/G-Research/fasttrackml/pkg/api/mlflow/service/run"
 	"github.com/G-Research/fasttrackml/pkg/database"
 	"github.com/G-Research/fasttrackml/pkg/repositories"
 	aimUI "github.com/G-Research/fasttrackml/pkg/ui/aim"
